Fixed error in assigning genes to intervals
This error appeared when I submitted a short list of intervals to an
enrichment analysis. In that case most genes do not overlap with any of
the complexity intervals. I had not accounted for this case, only for
cases where one end of the gene does not overlap any interval.